scripts/requirements.txt
matplotlib>=3.6
numpy>=1.23
Pillow>=9.0
imageio>=2.9
